GEODE-8980: Bump jedis from 3.5.2 to 3.6.0

+++ geode-apis-compatible-with-redis/src/distributedTest/java/org/apache/geode/redis/OutOfMemoryDUnitTest.java
@@ -188,7 +188,7 @@ public class OutOfMemoryDUnitTest {
   private void setRedisKeyAndValue(Jedis jedis, boolean withExpiration, String valueString,
       int keyNumber) {
     if (withExpiration) {
-      jedis.setex(FILLER_KEY + keyNumber, KEY_TTL_SECONDS, valueString);
+      jedis.setex(FILLER_KEY + keyNumber, (long) KEY_TTL_SECONDS, valueString);
     } else {
       jedis.set(FILLER_KEY + keyNumber, valueString);
     }
